backend/requirements.txt
groq==0.9.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
pdfplumber==0.11.0
requests==2.31.0
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
torch==2.3.0
transformers==4.40.2
scikit-learn==1.4.2
tqdm==4.66.4
